fix: 체크리스트 기반 조회 기능 구현
