+++ gg-admin-repo/src/main/java/gg/admin/repo/recruit/ApplicationAdminRepository.java
@@ -18,6 +18,19 @@ public interface ApplicationAdminRepository extends JpaRepository<Application, L
 	@EntityGraph(attributePaths = {"user", "applicationAnswers", "applicationAnswers.question"})
 	Page<Application> findByRecruitIdAndIsDeletedFalse(Long recruitId, Pageable pageable);
 
+	@EntityGraph(attributePaths = {"user", "applicationAnswers", "applicationAnswers.question",
+		"applicationAnswers.question.checkLists"})
+	@Query("SELECT a FROM Application a WHERE a.id IN " +
+		"(SELECT aa.application.id FROM ApplicationAnswerCheckList aa " +
+		"JOIN aa.checkList cl " +
+		"JOIN aa.application.recruit r " +
+		"WHERE r.id =:recruitId AND aa.question.id = :questionId AND cl.id IN :checkListIds) " +
+		"ORDER BY a.id DESC")
+	Page<Application> findAllByCheckList(
+		@Param("recruitId") Long recruitId,
+		@Param("questionId") Long questionId,
+		@Param("checkListIds") List<Long> checkListIds,
+		Pageable pageable);
 
 	@Query("SELECT a FROM Application a "
 		+ "JOIN FETCH a.user "
