+++ gg-recruit-api/src/test/java/gg/recruit/api/admin/controller/RecruitmentAdminControllerUnitTest.java
@@ -19,7 +19,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
@@ -165,7 +164,7 @@ class RecruitmentAdminControllerUnitTest {
 			//Assert
 			Assertions.assertThatThrownBy(
 					() -> recruitmentAdminController.getRecruitmentApplications(1L, null, "d,d,d", null,
-						PageRequest.of(1, 10, Sort.by(new Sort.Order(Sort.Direction.DESC, "id")))))
+						PageRequest.of(1, 10)))
 				.isInstanceOf(BusinessException.class);
 
 		}
@@ -176,7 +175,7 @@ class RecruitmentAdminControllerUnitTest {
 			//Arrange
 			//Act
 			recruitmentAdminController.getRecruitmentApplications(1L, null, "1,2,3", null,
-				PageRequest.of(1, 10, Sort.by(new Sort.Order(Sort.Direction.DESC, "id"))));
+				PageRequest.of(1, 10));
 
 			//Assert
 		}
